backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlmodel
google-generativeai>=0.8.0
python-dotenv
httpx
selenium
webdriver-manager
beautifulsoup4
pytest
pytest-asyncio
requests
